refactoring to Employer and Employee Controller and changes in routing yml
