+++ app/Resources/ShiftBundle/views/Shift/employer.html.twig
@@ -8,8 +8,8 @@
             </div>
             <div class="panel-body">
                 <ul>
-                    <li><a href="/shift/create">Create Shift</a></li>
-                    <li><a href="/shift/list">Open shift</a></li> 
+                    <li><a href="{{ url('createShift')}}">Create Shift</a></li>
+                    <li><a href="{{ url('employerListShifts')}}">Open shift</a></li> 
                     <li><a>Manage Profile</a></li>
                     <li><a>Invoices</a></li>
                 </ul>
