+++ app/Resources/ShiftBundle/views/Shift/listShifts.html.twig
@@ -8,7 +8,7 @@
                 {% for shift in shifts %}
                 <section>
                     <h3>
-                        <a href="{{ path('viewShift', {'id': shift.id}) }}">{{ shift.orgName()}}</a>
+                        <a href="{{ path('viewShift', {'id': shift.getId()}) }}">{{ shift.orgName()}}</a>
                     </h3>
 
                     <dl>
